@@ -28,13 +28,14 @@
     "@jswork/next-rollup-banner": "^1.0.2",
     "@jswork/react-markdown-props": "^1.0.1",
     "@rollup/plugin-image": "^2.1.1",
-    "@rollup/plugin-node-resolve": "^11.0.1",
+    "@rollup/plugin-node-resolve": "^15.0.1",
     "@rollup/plugin-replace": "^2.3.4",
+    "@rollup/plugin-commonjs": "^23.0.2",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@types/jest": "^27.0.3",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.8",
     "@types/react-test-renderer": "^17.0.0",
     "@types/styled-components": "^5.1.9",
     "@vitejs/plugin-react-refresh": "^1.3.1",
@@ -43,8 +44,8 @@
     "css-loader": "^5.0.1",
     "indent-string": "^4.0.0",
     "jest": "^27.4.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts-ts": "^3.1.0",
     "react-test-renderer": "^17.0.1",
     "release-it": "^15.2.0",
@@ -52,7 +53,6 @@
     "rollup": "^2.36.0",
     "rollup-plugin-auto-external": "^2.0.0",
     "rollup-plugin-banner": "^0.2.1",
-    "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-copy": "^3.3.0",
     "rollup-plugin-external-globals": "^0.6.1",
     "rollup-plugin-node-externals": "^2.2.0",
@@ -77,9 +77,5 @@
   "publishConfig": {
     "access": "<%= registry.access %>",
     "registry": "<%= registry.publish %>"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/afeiship/boilerplate-react-ts-component.git"
   }
 }
