@@ -3,9 +3,4 @@ import ReactDOM from 'react-dom';
 import '../../src/components/style.scss';
 import App from './app';
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+ReactDOM.createRoot(document.getElementById('root')!).render(<App />);
